Patch to fix potential build errors on Win: using CMake paths in Python.
The GrPython.cmake module substitutes CMake path strings directly into a
Python string that is then used to set the interpreter path for Python
scripts. Since Windows paths include backslashes, some elements of the
path can be interpreted as special characters or Unicode and cause
problems (path corrupted so conda-build doesn't substitute it out) or
build errors (invalid Unicode). This turns the python strings into raw
strings so that the backslashes are not interpreted.