Fix path corruption with unescaped backslashes on Windows with patch.
This turned out to be a genuine bug with GNU Radio's CMake scripts, so
this proper patch replaces a previous patch that was intended more as a
bandaid (and didn't work anyway).

The previous patch didn't work because the path variables in
gnuradio-runtime/lib/constants.cc.in are substituted *twice*, first in
gnuradio-runtime/lib/CMakeLists.txt and then in CMakeLists.txt. The
paths were fine in the first substitution, but not in the second
substitution because the variables lacked backslash escaping there.